src/requirements.txt
lxml
amba-event-stream
# pymongo
# ruamel.yaml
requests
spacy
spacytextblob